create_test_pdf.py
#!/usr/bin/env python3
"""
Simple script to create a PDF from markdown using reportlab
"""

try:
    from reportlab.pdfgen import canvas
    from reportlab.lib.pagesizes import letter
    from reportlab.lib.styles import getSampleStyleSheet
    from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer
    from reportlab.lib.units import inch
    
    def create_sudoku_pdf():
        # Read the markdown content
        with open('pdfs/sudoku-requirements.md', 'r', encoding='utf-8') as f:
            content = f.read()
        
        # Create PDF
        doc = SimpleDocTemplate(
            'pdfs/sudoku-requirements.pdf',
            pagesize=letter,
            rightMargin=72,
            leftMargin=72,
            topMargin=72,
            bottomMargin=18
        )
        
        styles = getSampleStyleSheet()
        story = []
        
        # Split content into sections
        lines = content.split('\n')
        
        for line in lines:
            line = line.strip()
            if not line:
                story.append(Spacer(1, 12))
                continue
                
            if line.startswith('# '):
                # Main heading
                story.append(Paragraph(line[2:], styles['Title']))
                story.append(Spacer(1, 12))
            elif line.startswith('## '):
                # Section heading
                story.append(Paragraph(line[3:], styles['Heading1']))
                story.append(Spacer(1, 12))
            elif line.startswith('### '):
                # Subsection heading
                story.append(Paragraph(line[4:], styles['Heading2']))
                story.append(Spacer(1, 6))
            elif line.startswith('- '):
                # Bullet point
                story.append(Paragraph("• " + line[2:], styles['Normal']))
                story.append(Spacer(1, 3))
            elif line.startswith('1. ') or line.startswith('2. ') or line.startswith('3. '):
                # Numbered list
                story.append(Paragraph(line, styles['Normal']))
                story.append(Spacer(1, 3))
            else:
                # Regular paragraph
                if line:
                    story.append(Paragraph(line, styles['Normal']))
                    story.append(Spacer(1, 6))
        
        doc.build(story)
        print("✅ PDF created: pdfs/sudoku-requirements.pdf")
        
    if __name__ == "__main__":
        create_sudoku_pdf()

except ImportError:
    print("⚠️ reportlab not available, creating simple text-based PDF alternative")
    
    # Fallback: Create a simple PDF with basic content
    def create_simple_pdf():
        from fpdf import FPDF
        
        class PDF(FPDF):
            def header(self):
                self.set_font('Arial', 'B', 15)
                self.cell(0, 10, 'Sudoku Web Application - Requirements', 0, 1, 'C')
                self.ln(10)
        
        # Read content
        with open('pdfs/sudoku-requirements.md', 'r', encoding='utf-8') as f:
            content = f.read()
        
        pdf = PDF()
        pdf.add_page()
        pdf.set_font('Arial', '', 10)
        
        # Add content (simplified)
        lines = content.split('\n')
        for line in lines[:50]:  # First 50 lines to avoid issues
            line = line.strip()
            if line and not line.startswith('#'):
                try:
                    pdf.cell(0, 6, line.encode('latin-1', 'ignore').decode('latin-1'), 0, 1)
                except:
                    pdf.cell(0, 6, 'Content line', 0, 1)
        
        pdf.output('pdfs/sudoku-requirements.pdf')
        print("✅ Simple PDF created: pdfs/sudoku-requirements.pdf")
    
    try:
        from fpdf import FPDF
        create_simple_pdf()
    except ImportError:
        print("❌ Neither reportlab nor fpdf available")
        print("Creating a mock PDF file for testing...")
        
        # Create a mock PDF with minimal content
        pdf_content = """%PDF-1.4
1 0 obj
<<
/Type /Catalog
/Pages 2 0 R
>>
endobj

2 0 obj
<<
/Type /Pages
/Kids [3 0 R]
/Count 1
>>
endobj

3 0 obj
<<
/Type /Page
/Parent 2 0 R
/MediaBox [0 0 612 792]
/Contents 4 0 R
/Resources <<
/Font <<
/F1 5 0 R
>>
>>
>>
endobj

4 0 obj
<<
/Length 200
>>
stream
BT
/F1 12 Tf
50 750 Td
(Sudoku Web Application - Requirements Document) Tj
0 -20 Td
(This is a requirements document for a Sudoku web application.) Tj
0 -20 Td
(The application should allow users to play interactive Sudoku puzzles.) Tj
0 -20 Td
(Features include: difficulty levels, hints, timer, and score tracking.) Tj
ET
endstream
endobj

5 0 obj
<<
/Type /Font
/Subtype /Type1
/BaseFont /Helvetica
>>
endobj

xref
0 6
0000000000 65535 f 
0000000010 00000 n 
0000000079 00000 n 
0000000136 00000 n 
0000000271 00000 n 
0000000519 00000 n 
trailer
<<
/Size 6
/Root 1 0 R
>>
startxref
575
%%EOF"""
        
        with open('pdfs/sudoku-requirements.pdf', 'w', encoding='latin-1') as f:
            f.write(pdf_content)
        print("✅ Mock PDF created for testing purposes")