requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
nltk==3.8.1
spacy==3.7.2
click==8.1.7
rich==13.7.0
python-docx==1.1.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# AI API Libraries
openai==1.40.6
azure-openai==1.0.0b8
ollama==0.3.1
httpx==0.27.0
aiohttp==3.9.1

# Configuration and utilities
pydantic==2.5.2
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2
jsonschema==4.20.0